Notify the user when an epoche finishes
When an epoche worth of images and there is less then 20 total epoches (so the console doesnt fill up with 2000 messages) there will be a notifaction in the terminal notifying the user that an epoche has finished so they can calculate about how long it will take to finish the captcha generation. The edited script was tested.